Add `LinuxPerf` extension for branch + instruction counts
This updates the core BenchmarkTools types to include `instructions` and
`branches` fields. If the extension is not available, these are `NaN`.

No support is included for measuring overhead or making judgements based
on these fields, but Serialization, Statistics, etc. are all supported
with their usual functionality for Trial / TrialEstimate / etc.